requirements.txt
tensorflow_gpu==1.5.0
numpy==1.17.4
matplotlib==3.1.2
opencv_python==4.1.2.30
Pillow==6.2.1

